refactor(FON-468): extract shared components and context layers
