+++ apps/client/src/features/nomination-files-table/components/cells/NominationFilesPrioritySelectorCell.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { useAffectationRow } from '../../context/files-affectations.context';
-import { PriorityBadge, PriorityBadgeList } from '@/components/shared/priorities/PriorityBadge';
+import { PriorityBadge, PriorityBadgeList } from '@/shared/components/priorities/PriorityBadge';
 import { DropdownSelectMultiple } from '@/shared/ui/DropdownSelectMultiple';
 import { PrioriteEnum } from '@/types/enums.types';
