+++ apps/client/src/features/observations/components/ObservationDetailsContent.tsx
@@ -3,8 +3,8 @@ import Card from '@codegouvfr/react-dsfr/Card';
 import { FormattedMessage } from 'react-intl';
 import { Link } from 'react-router';
 
-import { LolfiMagistratLink } from '@/components/shared/LolfiMagistratLink';
 import { useIsSg } from '@/features/auth/hooks/roles.hook';
+import { LolfiMagistratLink } from '@/shared/components/LolfiMagistratLink';
 import { TipTapEditor } from '@/shared/ui/tip-tap-editor';
 import type { FilesUploader } from '@/shared/ui/tip-tap-editor/extensions/editor-file-uploader';
 import { dateOnlyToDate } from '@/utils/date-only.util';
