+++ apps/client/src/pages/members/MemberListPage.tsx
@@ -1,6 +1,6 @@
-import { PageContentLayout } from '@/shared/ui/PageContentLayout';
 import { MemberList } from '@/features/members/components/MemberList';
 import { MemberListHeader } from '@/features/members/components/MemberListHeader';
+import { PageContentLayout } from '@/shared/ui/PageContentLayout';
 
 export function MemberListPage() {
   return (
