+++ apps/client/src/pages/reports/ReportListPage.tsx
@@ -6,7 +6,6 @@ import { defineMessage } from 'react-intl';
 import { generatePath, useParams } from 'react-router';
 
 import { ArchiveBannerPortal } from '@/components/shared/layouts/archived-banner/ArchiveBannerPortal';
-import { PriorityBadgeList } from '@/components/shared/priorities/PriorityBadge';
 import { ObservationLinks } from '@/features/observations/components/ObservationLinks';
 import { HeaderReportList } from '@/features/reports/components/HeaderReportList';
 import { ReportList } from '@/features/reports/components/ReportList';
@@ -14,6 +13,7 @@ import { ReportListViewToggle } from '@/features/reports/components/ReportListVi
 import { ReportsDnVueGenerale } from '@/features/reports/components/ReportsDnVueGenerale';
 import { ReportStateTag } from '@/features/reports/components/ReportStateTag';
 import { useReportListFocus } from '@/features/reports/hooks/useReportListFocus';
+import { PriorityBadgeList } from '@/shared/components/priorities/PriorityBadge';
 import { useDataTable, useQueryDataTableState } from '@/shared/ui/data-table';
 import { PageContentLayout } from '@/shared/ui/PageContentLayout';
 import {
