+++ apps/client/src/pages/spaces/member/MemberLayout.tsx
@@ -1,8 +1,8 @@
 import { Outlet, useLocation, useParams } from 'react-router';
 
-import { PageContentLayout } from '@/shared/ui/PageContentLayout';
 import { AuthGuard } from '@/features/auth/components/AuthGuard';
 import { AUTHORIZED_ROLES } from '@/features/auth/constants/authorized-roles.constants';
+import { PageContentLayout } from '@/shared/ui/PageContentLayout';
 
 export const MemberLayout = () => {
   const params = useParams();
