+++ apps/client/src/shared/components/LolfiMagistratLink.tsx
