+++ apps/client/src/shared/components/user-avatar/UserAvatar.tsx
