+++ apps/client/src/shared/components/user-avatar/index.ts
