+++ apps/client/src/shared/components/user-avatar/user-avatar.utils.ts
